Fix id_type to uint32_t in vendored darts.h; add legacy 64-bit OCD reader with full validation (#1373)
Inspection of OPENCCDARTS1 files produced by opencc_dict 1.3.2 on 64-bit
platforms confirms that the high 32 bits of every unit are zero, so the
logical double-array structure is identical to the 32-bit layout.  Switching
id_type from size_t to uint32_t is therefore safe, makes all future output
platform-independent, and keeps the vendored copy as close to upstream
darts-clone 0.32h as possible (one typedef change plus one added include).
A compatibility layer in DartsDict auto-detects old 64-bit files at load
time via a zero-bytes heuristic and reads them without altering the magic
string or on-disk format.

Detailed Changes:
- **deps/darts-clone-0.32h/include/darts.h**:
  - Add `#include <cstdint>`.
  - Change `typedef size_t id_type` → `typedef uint32_t id_type`; all other
    code is left untouched to stay in sync with upstream.
- **src/DartsDict.cpp**:
  - Add `LegacyUnit64` struct mirroring `DoubleArrayUnit` bit operations on
    `uint64_t` units.
  - Add `Legacy64ExactMatch` mirroring `exactMatchSearch` with the correct
    three-way `pos ^= offset ^ c` traversal and `pos ^ offset` leaf access.
  - Add `Legacy64PrefixSearch` mirroring `commonPrefixSearch`: continues
    traversal to `maxLen` even after the result buffer is full, always
    returning the true total match count so the caller's re-search path can
    allocate a correctly-sized buffer and retrieve the true longest match.
  - Add `ValidateLegacy64` mirroring `DoubleArrayImpl::validate` to check
    root sanity, offset bounds, and leaf value bounds before any traversal.
  - `NewFromFile`: read an 8-byte probe after the magic; bytes [4..7] all
    zero → old 64-bit layout (uint64_t dartsSize + 8-byte units, loaded into
    `legacyArray64` and validated); otherwise → 32-bit layout (uint32_t
    dartsSize + 4-byte units, seek back 4, existing path).
  - `MatchPrefix` legacy branch: restructure re-search to read `maxVal` from
    `rematchedResults`, not the original stack-allocated `results[64]` (was
    out-of-bounds and returned the wrong longest match).
  - `SerializeToFile`: write dartsSize as `uint32_t` instead of `size_t` so
    output is platform-independent; throw `InvalidFormat` early when called
    on a legacy-loaded dict whose `doubleArray` is nullptr.
- **src/DartsDictTest.cpp**:
  - Update `WriteMalformedDartsFile` / `CorruptDartsRootUnit` /
    `CorruptFirstDartsValue` to use `uint32_t` sizes.
  - Add `SerializeAsLegacy64` helper that zero-extends each 32-bit unit to
    64 bits to reproduce old build output.
  - Add `LoadsLegacy64bitFormat`: round-trip correctness via `TestDict` and
    an explicit Chinese-key lookup.
  - Add `RejectsInvalidLegacy64Root`: zeroed root unit must throw.
  - Add `Legacy64SerializeToFileThrows`: `SerializeToFile` on a
    legacy-loaded dict must throw `InvalidFormat` rather than segfault.
  - Add `Legacy64PrefixSearchBeyond64Matches`: 65-entry purely-nested dict
    round-tripped through legacy 64-bit OCD; `MatchPrefix` must return "v65"
    (the true longest match), not "v64" as it did when traversal stopped as
    soon as the result buffer was full.
  - Total DartsDictTest count: 12.

+++ deps/darts-clone-0.32h/include/darts.h
@@ -1,6 +1,7 @@
 #ifndef DARTS_H_
 #define DARTS_H_
 
+#include <cstdint>
 #include <cstdio>
 #include <cstring>
 #include <exception>
@@ -37,7 +38,7 @@ typedef int value_type;
 
 // The main structure of Darts-clone is an array of <DoubleArrayUnit>s, and the
 // unit type is actually a wrapper of <id_type>.
-typedef size_t id_type;
+typedef uint32_t id_type;
 
 // <progress_func_type> is the type of callback functions for reporting the
 // progress of building a dictionary. See also build() of <DoubleArray>.
